+++ src/Symfony/Component/Validator/Tests/Constraints/CidrValidatorTest.php
@@ -221,7 +221,6 @@ class CidrValidatorTest extends ConstraintValidatorTestCase
     {
         return [
             ['0.0.0.0/foobar'],
-            ['10.0.0.0/128'],
             ['123.45.67.178/aaa'],
             ['172.16.0.0//'],
             ['172.16.0.0/a/'],
@@ -240,6 +239,7 @@ class CidrValidatorTest extends ConstraintValidatorTestCase
     {
         return [
             ['10.0.0.0/24', Ip::V4, 10, 20],
+            ['10.0.0.0/128'],
             ['2001:0DB8:85A3:0000:0000:8A2E:0370:7334/24', Ip::V6, 10, 20],
         ];
     }